+++ src/fixActive.ts
@@ -1,0 +1,6 @@
+/*
+* Fix for :active.
+*/
+export default () => {
+  document.addEventListener('touchstart', () => {}, false);
+};