+++ src/initializeAppParams.ts
@@ -1,0 +1,10 @@
+/*
+ * Initializes app parameter with writing them in Window.
+ * Writes location params (search and location_hash), flag is_production
+ */
+export default () => {
+  window.search = location.search;
+  window.location_hash = location.hash;
+
+  window.is_production = process.env.NODE_ENV === 'production';
+};