+++ src/loadImages.ts
@@ -1,8 +1,13 @@
 import noop from './noop';
 
+/*
+ * Async wrapper for image.onload.
+ * @param {string} src Image src
+ * @param {VoidFunction} onLoaded Callback on image success load end
+ */
 export const loadImage = async (
   src: string,
-  onLoaded: () => void
+  onLoaded: VoidFunction
 ): Promise<void> =>
   new Promise<void>((resolve) => {
     const curImage = new Image();
@@ -14,9 +19,14 @@ export const loadImage = async (
     curImage.onerror = () => resolve();
   });
 
+/*
+ * Method for async load of multiple images
+ * @param {string[]} images Images src array
+ * @param {VoidFunction} onNextLoaded Callback on every image success load end
+ */
 export default async (
-  images: Array<string>,
-  onNextLoaded: () => void = noop
+  images: string[],
+  onNextLoaded: VoidFunction = noop
 ): Promise<void> => {
   await Promise.all(images.map((i) => loadImage(i, onNextLoaded)));
 };